+++ automatatutor-frontend/src/main/scala/com/automatatutor/snippet/MainSnippet.scala
@@ -4,6 +4,7 @@ import scala.Array.canBuildFrom
 import scala.xml._
 import com.automatatutor.lib._
 import com.automatatutor.model._
+import com.automatatutor.model.problems.WordsInGrammarProblem
 import com.automatatutor.renderer.{CourseRenderer, ProblemRenderer}
 import net.liftweb.common.Empty
 import net.liftweb.common.Full
@@ -131,9 +132,14 @@ class Mainsnippet {
 	val usersButton = SHtml.link("/users/index", () => (), <button type='button'>User List</button>)
 	val editProfilButton = SHtml.link("/user_mgt/change_password", () => (), <button type='button'>Change Password</button>)
   val becomeInstructorButton = SHtml.link("/main/become_instructor", () => {}, <button type='button'>Become Instructor</button>)
-	
-	if (user.isAdmin) return autogenButton ++ problemPoolButton ++ usersButton
-	if (user.isInstructor) return autogenButton ++ problemPoolButton ++ editProfilButton
+	val createGrammarWordsProbButton = SHtml.link("/main/problempool/create", () => {
+    val x = ProblemType.findByName("Grammar Words").head
+    println(x)
+    CurrentProblemTypeInCourse(x)
+  }, <button type='button'>Create a Words in Grammar Problem</button>)
+
+	if (user.isAdmin) return autogenButton ++ problemPoolButton ++ usersButton ++ createGrammarWordsProbButton
+	if (user.isInstructor) return autogenButton ++ problemPoolButton ++ editProfilButton ++ createGrammarWordsProbButton
 	return autogenButton ++ editProfilButton ++ becomeInstructorButton
   }
