+++ automatatutor-frontend/src/main/webapp/main/problempool/create.html
@@ -4,9 +4,8 @@
 <body class="lift:content_id=main">
 	<!--TODO 8/7/2020 do we need this?-->
 	<div id="main" class="lift:surround?with=default;at=content">
-	<!--			<lift:Coursesnippet.rendercreate>-->
-	<!--			</lift:Coursesnippet.rendercreate >-->
-	<h2>Testtt</h2>
+		<lift:Problempoolsnippet.rendercreate>
+		</lift:Problempoolsnippet.rendercreate >
 	</div>
   </body>
 </html>
