+++ automatatutor-frontend/src/main/webapp/templates-hidden/words-in-grammar-problem/create.html
@@ -21,10 +21,14 @@
 		<tr>
 			<td> Grammar: </td>
 			<td> <createform:grammarfield> </createform:grammarfield> </td>
-		</tr> 
+		</tr>
 		<tr>
-			<td> Short Description: </td>
-			<td> <createform:shortdescription> </createform:shortdescription> </td>
+			<td> Name: </td>
+			<td> <createform:namefield> </createform:namefield> </td>
+		</tr>
+		<tr>
+			<td> Description: </td>
+			<td> <createform:descriptionfield> </createform:descriptionfield> </td>
 		</tr>
 		<tr> 
 			<td> </td>
